feat(ForceSystemAlbum): Allow setting a default album type
Refactors the "Force System Album" feature from a simple toggle to a configurable option.

Previously, users were prompted to choose an album type every time. This change introduces a setting to select a default behavior, improving the user experience.

Users can now choose one of the following defaults:
- Ask every time (previous behavior)
- System Album
- System Documents
- QQ Album

A new settings entry is added to manage this preference, from which the feature can also be enabled or disabled.